requirements.txt
PyPDF2==3.0.1
google-genai
